+++ RELEASENOTES.md
@@ -9,6 +9,7 @@
   * Add `Player.onPlayWhenReadyChanged` with reasons.
   * Add `Player.onPlaybackStateChanged` and deprecate
     `Player.onPlayerStateChanged`.
+  * Deprecate and rename getPlaybackError to getPlayerError for consistency.
   * Make `MediaSourceEventListener.LoadEventInfo` and
     `MediaSourceEventListener.MediaLoadData` top-level classes.
   * Rename `MediaCodecRenderer.onOutputFormatChanged` to